feat: add 404-500 pages
- remove Helmet Package
- remove Overview Page
- remove Hero Page
- remove loading.tsx
- add Error404Section
- add Error500Section
- add Initial Structure of status page
- change the directorys of pages
