+++ src/app/404/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { Error404Section } from "@/components/04-templates";
+
+function Status404() {
+  return (
+    <div>
+      <Error404Section />
+    </div>
+  );
+}
+
+export default Status404;
