+++ src/app/500/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { Error500Section } from "@/components/04-templates/500Section";
+
+function Status500() {
+  return (
+    <div>
+      <Error500Section />
+    </div>
+  );
+}
+
+export default Status500;
