--- src/app/loading.tsx
@@ -1,5 +1,0 @@
-import Loader from "@/components/01-atoms/Loader";
-
-export default function Loading() {
-  return <Loader />
-}
