+++ src/components/01-atoms/index.tsx
@@ -1,3 +1,4 @@
 export * from "./ui";
+export * from "./LogoSign";
 export * from "./Loader";
 export * from "./Scrollbar";
