+++ src/components/03-organisms/index.tsx
@@ -1,0 +1,3 @@
+export * from "./CardMultiTasks";
+export * from "./CardTask";
+export * from "./SearchFiltersTasks";
