+++ src/components/04-templates/index.tsx
@@ -1,0 +1,2 @@
+export * from "./404Section";
+export * from "./500Section";
